refactor(piano-roll): split NUIPianoRollWidgets.cpp into per-class files
The 3,993-line TU held 8 widget classes. Mechanical split into one .cpp
per class plus PianoRollWidgetShared.h for the file-static helpers that
were shared across classes (isBlackKey, safeClamp*, beatToScreenX,
snap*X, getNoteLabel), now inline. NUIPianoRollWidgets.h is unchanged as
the umbrella header, so includers are untouched. No behavior change.
